+++ mountain/src/main.rs
@@ -62,8 +62,8 @@ use crate::systems::door::Parameters;
 use crate::systems::{
     building_artist, carousel, chair_artist, chair_framer, door, door_artist, frame_artist,
     frame_wiper, gate, gate_artist, global_computer, global_target_setter, lift_artist, log,
-    messenger, piste_adopter, planner, skiing_framer, target_scrubber, target_setter,
-    terrain_artist, tree_artist, window_artist,
+    messenger, piste_adopter, planner, selection_rasterizer, skiing_framer, target_scrubber,
+    target_setter, terrain_artist, tree_artist, window_artist,
 };
 use crate::utils::computer;
 use crate::widgets::{building_editor, main_menu, toaster};
@@ -573,7 +573,17 @@ impl EventHandler for Game {
             terrain_artist: &mut self.systems.terrain_artist,
             graphics,
         });
-        handlers::mode::handle(event, &self.bindings, &mut self.components.services.mode);
+        selection_rasterizer::run(selection_rasterizer::Parameters {
+            terrain: &self.components.terrain,
+            selection: &mut self.components.selection,
+            terrain_artist: &mut self.systems.terrain_artist,
+        });
+        handlers::mode::handle(
+            event,
+            &self.bindings,
+            &mut self.components.services.mode,
+            &mut self.components.selection,
+        );
 
         self.controllers
             .building_builder
